Updated docs to list the fixed issues with GEO location
working on photo stuff I did some research and relsoved that for 3rd
party systems image converted to base 64 string is easiest way

So I will add overloads for that I will comeplete that by thursday
night.  so lets meet at 8:30 on thursday.  For now please read the
updated documenation up to the end of the gelocation stuff.